various updates to use slpjs npm package
