+++ package.json
@@ -3,12 +3,13 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@material-ui/core": "^1.4.3",
-    "bitbox-cli": "^1.4.9",
+    "@material-ui/core": "^1.5.1",
+    "bitbox-cli": "^1.6.0",
     "qrcode.react": "^0.8.0",
     "react": "^16.4.2",
     "react-dom": "^16.4.2",
-    "react-scripts": "1.1.4"
+    "react-scripts": "1.1.4",
+    "slpjs": "^0.0.5"
   },
   "scripts": {
     "start": "react-scripts start",
